feat(homepage): add homepage with search and logo
1. Added a TwitterActionCreators file to setQuery as I type in the text
input field. I have yet to implement the async part of search (query
Twitter API).
2. Added a Logo component to display the TweetPedia logo
3. Added a SearchInput component (I want to change this into a container
later) which controls how we search. Eventually this will be shared with
the search-results page so I'll make it a container then.
4. Cleaned up the HTML a bit (too much nesting made CSS a bit annoying).
5. Updated the Express server to serve static files from /assets as well
as /public.

There's one last thing I'm uncertain about and it's how I implemented
the mapDispatchToProps function. When an action creator doesn't return a
new action, I don't want the dispatch to be fired. The way I've done it
is I've added a simple if inside nested functions within the
mapDispatchToProps function. I don't know if this is a good idea. It
technically doesn't mutate state, in fact it avoids it entirely. It
still remains deterministic because it's only based on input params so
it might be OK.